requirements.txt
black
uvicorn==0.17.6
fastapi==0.78.0
deta==1.1.0
jinja2==3.1.2
python-dotenv==0.20.0
passlib==1.7.4
python-multipart==0.0.5
python-jose==3.3.0